pyproject.toml
[tool.poetry]
name = "docker-influxdb-grafana-locust"
version = "0.1.0"
description = ""
authors = ["Bruno Tineli <bruno.tineli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
locustio = "=0.13.2"
locust-influx = "=0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry=0.12"]
build-backend = "poetry.masonry.api"
